+++ main.py
@@ -1,11 +1,25 @@
-from airflow import DAG
-from airflow.providers.docker.operators.docker import DockerOperator
-from airflow.operators.python import PythonOperator
-from datetime import datetime   
+import argparse
+import logging
+import os
+import sys
+import subprocess
+
+sys.path.append(os.path.abspath(os.path.dirname(__file__)))
+
+from utils.get_common_data import get_common_data
+from utils._enums import Solvers
 
 
-# Function for get_data task
-def get_data(**kwargs):
+def run_task(script, env_path=None, args=None):
+    """Run a Python script in a specific virtual environment."""
+    python_executable = f"{env_path}/bin/python3" if env_path else "python3"
+    command = [python_executable, script, *args]
+    if args:
+        command.extend(args)
+    subprocess.run(command, cwd="./", check=True)
+
+
+if __name__ == "__main__":
     parser = argparse.ArgumentParser(
         description="Runs tests of Causal Effect under Partial-Observability."
     )
@@ -15,99 +29,50 @@ def get_data(**kwargs):
     parser.add_argument('-v', '--verbose', action='store_true', help="Show solver logs")
     args = parser.parse_args()
 
-    # Creating a sample dictionary to pass
-    data = {'key1': 'value1', 'key2': 'value2'}
-    
-    # Pushing the dictionary to XCom
-    kwargs['ti'].xcom_push(key='data_dict', value=data)
+    common_data_path = "shared/common_data.json"
 
-# Function for solver_a task
-def solver_a(**kwargs):
-    # Pulling the dictionary from XCom (from the get_data task)
-    data = kwargs['ti'].xcom_pull(task_ids='get_data', key='data_dict')
-    
-    # Now you can use the data in solver_a
-    print(f"Data received in solver_a: {data}")
+    try:
+        if not args.verbose:
+            logging.getLogger().setLevel(logging.CRITICAL)
 
-# Function for solver_b task (example)
-def solver_b(**kwargs):
-    data = kwargs['ti'].xcom_pull(task_ids='solver_a', key='data_dict')
-    print(f"Data received in solver_b: {data}")
+        run_task(
+            "utils/input_processor.py",
+            # TODO: REVER O ENV DAQUI 
+            env_path="venv_bcause",
+            args=["--output", common_data_path, "--input", args.file_path]
+        )
 
-# Define the DAG
-dag = DAG(
-    'solver_pipeline_with_xcom',
-    start_date=datetime(2023, 11, 14),
-    schedule_interval=None,  # No periodic schedule
-)
+        data = get_common_data(common_data_path)
 
-# Define tasks
-get_data_task = PythonOperator(
-    task_id='get_data',
-    python_callable=get_data,
-    provide_context=True,  # Ensure context is passed to the function
-    dag=dag,
-)
+        if Solvers.DOWHY.value in data["solvers"]:
+            run_task(
+                "src/solvers/dowhy_solver.py",
+                env_path="venv_dowhy",
+                args=["--common_data", common_data_path]
+            )
 
-solver_a_task = PythonOperator(
-    task_id='solver_a',
-    python_callable=solver_a,
-    provide_context=True,
-    dag=dag,
-)
+        if Solvers.BCAUSE.value in data['solvers']:
+            run_task(
+                "src/solvers/bcause_solver.py",
+                env_path="venv_bcause",
+                args=["--common_data", common_data_path]
+            )
 
-# Set task dependencies
-get_data_task >> solver_a_task
+        if Solvers.AUTOBOUNDS.value in data["solvers"]:
+            run_task(
+                "src/solvers/autobounds_solver.py",
+                env_path="venv_autobounds",
+                args=["--common_data", common_data_path]
+            )
 
-
-# dag = DAG(
-#     'solver_pipeline',
-#     start_date=datetime(2023, 11, 14),
-#     schedule_interval=None,
-# )
+        if Solvers.LCN.value in data["solvers"]:
+            run_task(
+                "src/solvers/lcn_solver.py",
+                env_path="venv_lcn",
+                args=["--common_data", common_data_path]
+            )
 
 
-# get_input_task = DockerOperator(
-#     task_id='run_get_input',
-#     image='get_input_image',
-#     api_version='auto',
-#     auto_remove=True,
-#     dag=dag,
-# )
-
-# solver_dowhy_task = DockerOperator(
-#     task_id='run_solver_dowhy',
-#     image='solver_dowhy_image',
-#     api_version='auto',
-#     auto_remove=True,
-#     dag=dag,
-# )
-
-
-# solver_bcause_task = DockerOperator(
-#     task_id='run_solver_bcause',
-#     image='solver_bcause_image',
-#     api_version='auto',
-#     auto_remove=True,
-#     dag=dag,
-# )
-
-
-# solver_lcn_task = DockerOperator(
-#     task_id='run_solver_lcn',
-#     image='solver_lcn_image',
-#     api_version='auto',
-#     auto_remove=True,
-#     dag=dag,
-# )
-
-# solver_autobounds_task = DockerOperator(
-#     task_id='run_solver_autobounds',
-#     image='solver_autobounds_image',
-#     api_version='auto',
-#     auto_remove=True,
-#     dag=dag,
-# )
-
-
-# solver_dowhy_task >> solver_bcause_task >> solver_autobounds_task >> solver_lcn_task
+    except Exception as e:
+        print(f"{type(e).__module__}.{type(e).__name__}: {e}")
+
